Done
sorry for the late submission for the second project.
I meant to send this in a lot sooner but wanted to improve upon it and it was sent to the back burner.
I got everything i wanted to implement in except for scoring and highscoring.

the code provided is much more difficult to read and go around than i  though.

TBH i dont want to think about this late over dued project anylonger so im submitting it even without completely finishing it

things implemented:
collision
music
game start screen
game over screen

to do:
scoring,
highscoring

maybe make it into an infinite scolling and dodging game like the chrome dino game complete with gravity and enemies